Change #1.0
- Worked on square class, and used an array of squares filled different
colors as a color selector. When mouse is clicked inside one of the
squares, it changes the sand to that color (currently changes all
existing sand, need to change to all NEW sand)
- Sand now falls at a greater velocity, and if the sand reaches the
bottom, it stops falling (changed from == height - 2 to >=, as this
would better allow it to detect when sand had reached the bottom).
- If sand is near another particle, it will be shifted up 2. However,
this is not working well, as the particles “fly” for the most part,
never being more than the needed distance not to be “touching,” as
multiple move up at once.